requirements.txt
decorator
flask
requests
toolforge
python-stdnum
